+++ resources/release-notes.md
@@ -2,7 +2,9 @@
 
 - Improve handling of "croppable" data (inheriting from `mag.mixin.Croppable`) with utility methods `mustBeTimeFilter` and `convertToTimeSubscript`
 - Do not use "Primary" or "Secondary" in sensor event labels
+- Single-source definition of empty event and science event timetables
 - Fix [#43](https://github.com/ImperialCollegeLondon/MAG-Data-Visualization-Toolbox/issues/43)
+- Fix issue with `mag.Instrument/Events` not being part of deep copy
 
 # GitHub